+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 3.7.0
 
 * added `True-Client-IP` header to receive the real visitor IP (Cloudflare)
+* added new bot list by Atmire (https://github.com/atmire/COUNTER-Robots/blob/master/COUNTER_Robots_list.json)
+* added script to update and clean up the User-Agent blacklist
 * fixed grouping referrers with path in URL
 * updated User-Agent blacklist
 * updated referrer blacklist